Update particle tracks
Check files exist and that more than one set of coordinates are
available for line joining

+++ S8_Display_Particle_Tracks.m
@@ -6,6 +6,8 @@
 
 function S8_Display_Particle_Tracks(XLS)
 
+w = waitbar(0,'Reading MAT and XLS data');
+
 % Set the base pathname for the current machine
 setbasepath;
 
@@ -21,6 +23,8 @@ if ~exist([basepath,[basepath,expt.tracking.tracks]]), mkdir([basepath,expt.trac
 
 %% Write the annotated images
 for s = 1:length(sheets),
+    
+    waitbar(s/length(sheets),w,['Reading sheet: ',sheets{s}(1:length(sheets{s})-1)]);
     
     if ~strcmp(sheets{s},'Sheet1') & ~strcmp(sheets{s},'Sheet2') & ~strcmp(sheets{s},'Sheet3') & ~strcmp(sheets{s},'Mean') & ~strcmp(sheets{s},'SD') & ~strcmp(sheets{s},'Number'),
         
@@ -52,36 +56,44 @@ for s = 1:length(sheets),
                     framenumber,...
                     expt.fad.FAD_type_low);
                 
-                % Load and prep the image data
-                im = imread(filename);
-                im = repmat(im,[1 1 3]);
-                im = imresize(im,SCALE);
-                
-                for p = unique(data(data(:,1) == trackingtimes(t), 2))'
-                  
-                    % Get the coordinates of the current particle
-                    coordinates = data((data(:,1) == trackingtimes(t)) & (data(:,2) == p) & isfinite(data(:,4)), 4:5);
+                if exist(filename),
                     
-                    % Add the X marker showing initial position
-                    markerInserter = vision.MarkerInserter ('Shape','X-mark','Size',10,'BorderColor','Custom','CustomBorderColor',[255 0 0]);
-                    marker = SCALE*int32(coordinates(1,:));
-                    im = step(markerInserter, im, marker);
+                    % Load and prep the image data
+                    im = imread(filename);
+                    im = repmat(im,[1 1 3]);
+                    im = imresize(im,SCALE);
                     
-                    % Add the O markers showing subsequent positions
-                    markerInserter = vision.MarkerInserter ('Shape','Circle','Size',2,'Fill',1,'FillColor','Custom','CustomFillColor',[0 0 255]);
-                    marker = SCALE*int32(coordinates(2:size(coordinates,1),:));
-                    im = step(markerInserter, im, marker);
+                    for p = unique(data(data(:,1) == trackingtimes(t), 2))'
+                        
+                        % Get the coordinates of the current particle
+                        coordinates = data((data(:,1) == trackingtimes(t)) & (data(:,2) == p) & isfinite(data(:,4)), 4:5);
+                        
+                        % Add the X marker showing initial position
+                        markerInserter = vision.MarkerInserter ('Shape','X-mark','Size',10,'BorderColor','Custom','CustomBorderColor',[255 0 0]);
+                        marker = SCALE*int32(coordinates(1,:));
+                        im = step(markerInserter, im, marker);
+                        
+                        % Add the O markers showing subsequent positions
+                        markerInserter = vision.MarkerInserter ('Shape','Circle','Size',2,'Fill',1,'FillColor','Custom','CustomFillColor',[0 0 255]);
+                        marker = SCALE*int32(coordinates(2:size(coordinates,1),:));
+                        im = step(markerInserter, im, marker);
+                        
+                        if size(coordinates,1) > 1,
+                            
+                            % Add the line tracks joining the positions
+                            shapeInserter = vision.ShapeInserter('Shape','Lines','BorderColor','Custom','CustomBorderColor',[255 255 0]);
+                            shape = SCALE * reshape(coordinates',1,2*size(coordinates,1));
+                            im = step(shapeInserter, im, int32(shape));
+                            
+                        end
+                        
+                    end
                     
-                    % Add the line tracks joining the positions
-                    shapeInserter = vision.ShapeInserter('Shape','Lines','BorderColor','Custom','CustomBorderColor',[255 255 0]);
-                    shape = SCALE * reshape(coordinates',1,2*size(coordinates,1));
-                    im = step(shapeInserter, im, int32(shape));
+                    % Write the image
+                    filename = sprintf('%s%s%s%s%.1f%s%s',basepath,expt.tracking.tracks,sheets{s},'_',trackingtimes(t),'_min',expt.fad.FAD_type_low);
+                    imwrite(im,filename);
                     
                 end
-                
-                % Write the image
-                filename = sprintf('%s%s%s%s%.1f%s%s',basepath,expt.tracking.tracks,sheets{s},'_',trackingtimes(t),'_min',expt.fad.FAD_type_low);
-                imwrite(im,filename);
                 
             end
             
@@ -90,3 +102,5 @@ for s = 1:length(sheets),
     end
     
 end
+
+close(w)
